global: fix model and rules structure
* conflicting rules moved to videos (contributors, access)

* new layer of base model created for video and project which
  contains common fields for both
